reduce object churn
Cache the connection and client in the server object. If a client
is cleanly disconnected we don't need to tear it all down. This will
mean any negotiations made at the protocol level will hold when the
client reconnected. However the post_connect callback at the
application layer will still be called.